Changes to be committed: 	deleted:    binary/RecView160912x64CUDA.exe 	deleted:    binary/RecView160912x86CUDA.exe 	new file:   binary/RecView160920x64CUDA.exe 	new file:   binary/RecView160920x86CUDA.exe 	modified:   source/clReconst.cpp 	modified:   source/cudaReconst.cu 	modified:   source/cudaReconst.h 	modified:   source/cudaReconstHost.cpp 	modified:   source/projx64.lst

+++ source/clReconst.cpp
@@ -544,7 +544,7 @@ void CLReconstHost(RECONST_INFO* ri, int idev, bool bReport) {
 		//	break;
 		//}
 
-		float theta = (ri->fdeg[i] + ri->fTiltAngle) * (float)DEG_TO_RAD;
+		double theta = (ri->fdeg[i] + ri->fTiltAngle) * DEG_TO_RAD;
 		const float fcos = (float)(cos(theta) * DBPT_GINTP);
 		const float fsin = (float)(-sin(theta) * DBPT_GINTP);
 		const float fcenter = (float)((ixdimh + center - (int)(center)) * DBPT_GINTP);
